applied the patch from Jean-Pierre CHARRAS fixing C locale issue; replaced previous fgets() based FILE_LINE_READER::ReadLine() implementation with the new Dick Hollenbeck's one based on fgetc(); replaced size_t back to integer

+++ common/filter_reader.cpp
@@ -27,9 +27,9 @@
 #include <richio.h>
 #include <filter_reader.h>
 
-size_t FILTER_READER::ReadLine() throw( IO_ERROR )
+unsigned FILTER_READER::ReadLine() throw( IO_ERROR )
 {
-    size_t ret;
+    unsigned ret;
 
     while( ( ret = reader.ReadLine() ) != 0 )
     {
